refactor: const-generic device sharing
