+++ crates/machine_x68k/src/bus.rs
@@ -35,7 +35,7 @@ use device::{
     sasi::X68kSasiHdc,
     scsi::Mb89352Spc,
     sprite_x68k::SpriteX68k,
-    upd72065_fdc::Upd72065Fdc,
+    upd765a_fdc::{UPD765_PLATFORM_X68K, Upd765aFdc},
     video_controller_x68k::VideoControllerX68k,
     ym3802::{YM3802_CLKM_HZ, Ym3802},
     z8530::Z8530,
@@ -242,7 +242,7 @@ pub struct X68kBus<T: TraceSink = NoTrace> {
     rtc: Rp5c15Rtc,
     keyboard: KeyboardX68k,
     dmac: Hd63450Dmac,
-    fdc: Upd72065Fdc,
+    fdc: Upd765aFdc<UPD765_PLATFORM_X68K>,
     fdd: FddX68k,
     hdc: X68kSasiHdc,
     spc: Mb89352Spc,
@@ -358,7 +358,7 @@ impl<T: TraceSink> X68kBus<T> {
             rtc: Rp5c15Rtc::new(),
             keyboard: KeyboardX68k::new(),
             dmac: Hd63450Dmac::new(),
-            fdc: Upd72065Fdc::new(),
+            fdc: Upd765aFdc::new(),
             fdd: FddX68k::new(),
             hdc: X68kSasiHdc::new(cpu_clock_hz),
             spc: Mb89352Spc::new(cpu_clock_hz),
